pyproject.toml
[project]
name = "image-duplicates"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "imagehash>=4.3.2",
    "imageio>=2.37.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "streamlit>=1.47.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
]
